AWS-211 implement get classic libraries
This commit includes the get end point for obtaining classic libraries.
A suite of tests have been included with stub data. Only a 200 is expected,
with a specific set of content. Any error messages or other types of responses
are handled in a generic way.

A database to store the cookies has been adopted instead of utilising the
myads data store. There are a few reasons for this, specifically for maintainance
purposes, and you can read more about it in the JIRA ticket.

All tests use testing.postgresql, and hard code the database within the
tests to avoid any issues.

The database test class should be slimlined into a base.py file, and then the
database can be included within the authentication end point.