Feature Issue #11: ADS2.0 Zotero Export
Added the export feature for Zotero libraries. It generates a .zip
files, that contains each library as a .bib file. Tags are added as
the 'keyword' parameter, and notes under the 'notes' parameter.

MongoDB flat files must be included for this end point to work, and
they must exist on S3. The only difference being they have a different
extension to the other files, '.tags.json'.